pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mirnaprotpred"
version = "0.1.6"
description = "A package for miRNA protein prediction with SeqFinder and validator modules"
readme = "README.md"
requires-python = ">=3.13"
license = {text = "MIT"}
authors = [
    {name = "Sudipta Sardar", email = "sudipta@pusan.ac.kr"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "pandas>=2.3.3",
    "openpyxl>=3.1.5",
    "biopython>=1.86",
    "pyfiglet>=1.0.4",
    "requests>=2.32.5",
    "viennarna>=2.7.1",
]

[project.urls]
Homepage = "https://github.com/somenath-combio/mirnaprotpred"

[project.scripts]
SeqFinder = "mirnaprotpred.SeqFinder.seqfinder:cli"
validator = "mirnaprotpred.validator.validator:cli"
